Improve lock behavior for ChunkLock.
A combination of engines is necessary to properly handle all lock
situations.

An overwrite is used instead of an update when trying to acquire the
lock initially, since some attributes of the lock object will be
seen as "unchanged" since the last load, even though the object in
dynamo was deleted since the last load.  Because cross-object changes
are not tracked, it lead to the worker attribute not being persisted.

The condition prevents an overwrite from removing any existing locks;
now, a lock can only be acquired IFF there is no exising lock;
an existing lock past the timeout will first be deleted by the acquire
call, and then subsequent acquire call will attempt to obtain the lock.

Release still uses an atomic delete, since we should only ever delete
the exact persisted version of the lock.  Deleting anything else could
race with other changes.  The same behavior is used with renew, since
a lock should never renew against unexpected changes (DynamoDB should
match exactly as expected to change the time field).